Build with our own version of emscripten, don't do any optimisation. Todo.
Build with our own version of emscripten (the last, the provided one is very old).
Don't do any optimisation. This might be added again when everything starts to work, but it doesn't make much sense in a node environment.

Also, the version of graphviz used is rather old. Tried the newer one, but this code is patched, and the existing patches don't apply.
See if the patches can be applied to the newer version without going too deep.

Also, this version only support text-based image formats, not any binary formats.
Try to get that in.

The result is now a "rather big" JavaScript file. Node has no problems with it. WebStorm does. TextWrangler handles it with ease. By adding binary format support, it will probably become humongous.

Remind yourself that the reason for doing this with JavaScript is easy deployment on a Node provider, without the need for a Linux VM!