Reworked Makefile to the new situation. Not done.
This "works", but is incomplete, and too coarse.

In complete, because line 22 is commented out: we are doing nothing for gvc now.
This gave an error about not finding a VAR GVLIBDIR. This wasn't a problem in the viz.js version, so why is it now?
Probably, we fall through some other test, and we didn't before?

VAR names a bit more clear for me.

We now need to bring the graphviz source on par. The distribution tarbal has a lot of generated stuff in it, that isn't in the repo.
For more info, see the README of the Toryt/graphviz emscripten directory.
The target $(GRAPHVIZ_DIR)/Makefile does this: autogen.sh + configure are documented to be needed.
But there is something extra, in the tarbal, that we don't get. Some .h files and .c files (htmlparse.h) are also generated, and not by any of these 2 commands. They are generated in make commands. Which command we have to do exactly, to get only this, is not clear to me. It might be so that this is possible on the top level at all. make all does "enough", but probably way too much. This might be resolved best by bringing each of our make commands, for the parts, into the graphviz code tree.

I needed to add a reference to -I../ortho in 2 places (dotgen, neatogen). We didn't need this before. Unclear why we need this now.

Changed stuff to work with new position in dir structure (-I../../.. --> -I../../../..). This resolved a problem where a config.h file could not be located. This file probably should move to the grahviz project too?

Cleaning and clobbering: clobber does away too much? Not enough?